chunking/rag_app/requirements.txt
pinecone>=3.0.0
sentence-transformers
python-dotenv


